% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tables.R
\name{add_contents_sheet}
\alias{add_contents_sheet}
\title{Add contents sheet to give given workbook based on tibble of sheet titles}
\usage{
add_contents_sheet(wb, contents, contents_title)
}
\arguments{
\item{wb}{Workbook object}

\item{contents}{Contents}

\item{contents_title}{Character string of title to include on contents page.}
}
\description{
Add contents sheet to give given workbook based on tibble of sheet titles
}
