% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tables.R
\name{add_data_sheet}
\alias{add_data_sheet}
\title{Create table sheet with header wording, tables and notes}
\usage{
add_data_sheet(
  wb,
  sheet_name,
  sheet_title,
  sheet_tables,
  notes_list,
  note_mapping,
  n_tables,
  notes_start,
  use_low,
  table_headings
)
}
\arguments{
\item{wb}{Workbook object}

\item{sheet_name}{Sheet name}

\item{sheet_title}{Sheet title}

\item{sheet_tables}{Sheet tables}

\item{notes_list}{List of notes}

\item{note_mapping}{Notes to included against each table}

\item{n_tables}{Number of tables}

\item{notes_start}{Row that notes start on}

\item{use_low}{Does the page use low for non-sensitive values}

\item{table_headings}{Use subheadings for all tables}
}
\description{
Create table sheet with header wording, tables and notes
}
