% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tables.R
\name{add_data_sheets}
\alias{add_data_sheets}
\title{Add Data Sheet}
\usage{
add_data_sheets(wb, table_layout, notes_list)
}
\arguments{
\item{wb}{Workbook object}

\item{table_layout}{Table layout generated by metadata function}

\item{notes_list}{List of notes}
}
\description{
Add Data Sheet
}
