% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tables.R
\name{add_data_table}
\alias{add_data_table}
\title{Add Data Table}
\usage{
add_data_table(
  wb,
  sheet_name,
  sheet_title,
  table,
  start_row,
  end_row,
  header_rows,
  table_name
)
}
\arguments{
\item{wb}{Workbook object}

\item{sheet_name}{Sheet name}

\item{sheet_title}{Sheet title}

\item{table}{Table to be added}

\item{start_row}{Sheet row on which table starts}

\item{end_row}{Sheet row on which table ends}

\item{header_rows}{Number of header rows}

\item{table_name}{Name of table}
}
\description{
Add Data Table
}
