% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tables.R
\name{add_data_tables}
\alias{add_data_tables}
\title{Add data tables to existing worksheet}
\usage{
add_data_tables(
  wb,
  sheet_name,
  sheet_title,
  tables,
  header_rows,
  table_headings
)
}
\arguments{
\item{wb}{Workbook object}

\item{sheet_name}{Sheet name}

\item{sheet_title}{Sheet title}

\item{tables}{Tables to be included on sheet}

\item{header_rows}{Number of header rows}

\item{table_headings}{Use subheadings for all tables}
}
\description{
Add data tables to existing worksheet
}
