% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tables.R
\name{add_notes_sheet}
\alias{add_notes_sheet}
\title{Add notes sheet to given workbook}
\usage{
add_notes_sheet(wb, contents, notes_list)
}
\arguments{
\item{wb}{Workbook object}

\item{contents}{Contents}

\item{notes_list}{List of notes to include in publication}
}
\description{
Add notes sheet to given workbook
}
