% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_metadata.R
\name{add_sheet_to_metadata}
\alias{add_sheet_to_metadata}
\title{Instructs metadata to add tables to a given worksheet}
\usage{
add_sheet_to_metadata(
  metadata,
  sheet_name,
  sheet_title,
  table_names,
  table_notes
)
}
\arguments{
\item{metadata}{Metadata object}

\item{sheet_name}{Sheet name}

\item{sheet_title}{Sheet title}

\item{table_names}{List of table names}

\item{table_notes}{List of table notes}
}
\description{
Instructs metadata to add tables to a given worksheet
}
