% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_metadata.R
\name{add_tables_to_metadata}
\alias{add_tables_to_metadata}
\title{Add table metadata for all tables}
\usage{
add_tables_to_metadata(metadata, table_list)
}
\arguments{
\item{metadata}{Metadata object}

\item{table_list}{List of tables}
}
\description{
Add table metadata for all tables
}
