% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_metadata.R
\name{create_metadata}
\alias{create_metadata}
\title{Initialise Metadata Object}
\usage{
create_metadata()
}
\description{
Creates an empty metadata object which can be added to with
\code{xlsss::add_sheet_to_metadata}
}
