% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_metadata.R
\name{create_table_layout}
\alias{create_table_layout}
\title{Combine Metadata and Tables to Create Layout}
\usage{
create_table_layout(metadata, table_data)
}
\arguments{
\item{metadata}{Metadata object}

\item{table_data}{Tibble of tables. Must include columns: name, table and title.
\code{name} column must match the table names specified in the metadata object.
\code{table} column contains the tables to be outputted to excel
\code{title} column is only used where more than one table is included on a sheet
and is the sub title to be printed above the table.}
}
\description{
Takes a metadata object and adds specified table_data to create a table layout to
be passed to \verb{r xlsss::make_output_tables}
}
