% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_template.R
\name{create_template_output}
\alias{create_template_output}
\title{Create Template Script to Output Tables}
\usage{
create_template_output(filename = "template")
}
\arguments{
\item{filename}{File to output template script to}
}
\description{
Create Template Script to Output Tables
}
