% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tables.R
\name{format_columns}
\alias{format_columns}
\title{Format Columns}
\usage{
format_columns(wb, sheet_name, table, column, start_row, end_row)
}
\arguments{
\item{wb}{Workbook object}

\item{sheet_name}{Name of sheet}

\item{table}{Table to be formatted}

\item{column}{Column to be formatted}

\item{start_row}{Row to start formatting}

\item{end_row}{Row to end formatting}
}
\description{
Format Columns
}
