% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tables.R
\name{format_rows}
\alias{format_rows}
\title{Format Rows}
\usage{
format_rows(wb, sheet_name, table, table_row, sheet_row, start_col, end_col)
}
\arguments{
\item{wb}{Workbook object}

\item{sheet_name}{Name of sheet}

\item{table}{Table to be formatted}

\item{table_row}{Row of table to be formatted}

\item{sheet_row}{Row on sheet to be formatted}

\item{start_col}{Start column of formatting}

\item{end_col}{End row of formatting}
}
\description{
Format Rows
}
