% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_metadata.R
\name{generate_table_metadata}
\alias{generate_table_metadata}
\title{Generate metadata for a single sheet}
\usage{
generate_table_metadata(
  table_names,
  table_list,
  padding_rows_multi = 2,
  padding_rows_single = 1
)
}
\arguments{
\item{table_names}{Table name}

\item{table_list}{List of tables}

\item{padding_rows_multi}{Row gap for sheet with multiple tables}

\item{padding_rows_single}{Row gap for sheet with single table}
}
\description{
Generate metadata for a single sheet
}
