% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tables.R
\name{make_contents_table}
\alias{make_contents_table}
\title{Create tibble of sheet titles based on publication date}
\usage{
make_contents_table(publication_date, table_layout)
}
\arguments{
\item{publication_date}{Publication date}

\item{table_layout}{Table layout object created by metadata functions}
}
\description{
Create tibble of sheet titles based on publication date
}
