% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tables.R
\name{make_output_tables}
\alias{make_output_tables}
\title{Create excel tables}
\usage{
make_output_tables(
  metadata,
  table_data,
  notes_list,
  contents_title,
  tweak = TRUE,
  use_low = FALSE,
  table_headings = FALSE
)
}
\arguments{
\item{metadata}{metadata object created by metadata functions}

\item{table_data}{Tibble of tables. Must include columns: name, table and title.
\code{name} column must match the table names specified in the metadata object.
\code{table} column contains the tables to be outputted to excel
\code{title} column is only used where more than one table is included on a sheet
and is the subtitle to be printed above the table.}

\item{notes_list}{List of notes in publication}

\item{contents_title}{Title of contents page}

\item{tweak}{Should format tweaking be applied}

\item{use_low}{Does the wb use low for non-sensitive values}

\item{table_headings}{Use subheadings for all tables}
}
\description{
Create excel tables
}
