% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tables.R
\name{negative_to_c}
\alias{negative_to_c}
\title{Negative to "c"}
\usage{
negative_to_c(
  wb,
  sheet_name,
  table,
  column,
  row,
  start_row,
  end_row,
  header_rows
)
}
\arguments{
\item{wb}{Workbook object}

\item{sheet_name}{Sheet name}

\item{table}{Table to be formatted}

\item{column}{Column to format}

\item{row}{Row to format}

\item{start_row}{Row to start formatting}

\item{end_row}{Row to end formatting}

\item{header_rows}{Number of header rows}
}
\description{
Negative to "c"
}
