% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_metadata.R
\name{table_list_to_tibble}
\alias{table_list_to_tibble}
\title{Convert List of Tables to table_data}
\usage{
table_list_to_tibble(table_list)
}
\arguments{
\item{table_list}{List of tables to be converted to table_data}
}
\description{
Helper function for quickly turning named list of tables into a
table_data object for use in the \code{xlsss::create_table_layout} function.
}
