% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tables.R
\name{tweak_formatting}
\alias{tweak_formatting}
\title{Specific changes to individual sheet formatting}
\usage{
tweak_formatting(wb)
}
\arguments{
\item{wb}{Workbook object}
}
\description{
Specific changes to individual sheet formatting
}
