% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsss-package.R
\docType{package}
\name{xlsss-package}
\alias{xlsss}
\alias{xlsss-package}
\title{xlsss: Automate Accessible Table Production for Social Security Scotland}
\description{
Automates the process of outputting statistical tables to excel in a format that complies with accessibility requirements.
}
\author{
\strong{Maintainer}: Izaak Jephson \email{first.last@example.com}

}
\keyword{internal}
