% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tables.R
\name{add_data_sheets}
\alias{add_data_sheets}
\title{Add Data Sheet}
\usage{
add_data_sheets(
  wb,
  table_layout,
  notes_list,
  use_low = FALSE,
  table_headings = FALSE
)
}
\arguments{
\item{wb}{Workbook object}

\item{table_layout}{Table layout generated by metadata function}

\item{notes_list}{List of notes}

\item{use_low}{Does the wb use low for non-sensitive values}

\item{table_headings}{Use subheadings for all tables}
}
\description{
Add Data Sheet
}
