% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_metadata.R
\name{add_sheet_to_metadata}
\alias{add_sheet_to_metadata}
\title{Add Sheet to Metadata Object}
\usage{
add_sheet_to_metadata(
  metadata,
  sheet_name,
  sheet_title,
  table_names,
  table_notes
)
}
\arguments{
\item{metadata}{Metadata object}

\item{sheet_name}{Name to be displayed on sheet tab}

\item{sheet_title}{Title to be show in bold in cell A1 on sheet}

\item{table_names}{Character vector of table names to include on this sheet. Names must match those in the table_list object}

\item{table_notes}{Numeric vector of notes to be included underneath tables on this sheet}
}
\description{
Adds a new sheet to a metadata object.
}
