% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tables.R
\name{make_output_tables}
\alias{make_output_tables}
\title{Create excel tables}
\usage{
make_output_tables(
  table_layout,
  notes_list,
  publication_date,
  contents_title,
  workbook_filename
)
}
\arguments{
\item{table_layout}{Table layout object created by metadata functions}

\item{notes_list}{List of notes in publication}

\item{publication_date}{Publication date}

\item{contents_title}{Title of contents page}

\item{workbook_filename}{Filename to export workbook to}
}
\description{
Create excel tables
}
